Implement Credential Caching
This introduces a socket based credential cache to allow caching of keys
so that users do not need to go through the OIDC flow multiple times
for batch operations (e.g. rebases).

Credentials are keyed to git working directories so that different repos can
cache different identities, and credentials are never directly stored to disk.

Signed-off-by: Billy Lynch <billy@chainguard.dev>