Only include leaf certificate in signature.
Turns out cert ordering was still non-deterministic, so restricting the
certs included to just the leaf cert (which is probably okay in most
cases?)

We might add support back for this later, but for now this should be
okay.

Also fix e2e test so failing validation actually causes a non-zero exit
code.

Signed-off-by: Billy Lynch <billy@chainguard.dev>