e2e: select checkout ref based on event type.
This change adds logic to select the ref to checkout for events.

By default pull_request_target will use the base branch as the
target since it was originally intended for trusted workloads.
However, we need to use this to have access to the OIDC creds
for the e2e tests, so insert our own logic here.
This is effectively a ternary of the form ${{ <condition> <true> || <false> }}.
See https://docs.github.com/en/actions/learn-github-actions/expressions for more details.

Signed-off-by: Billy Lynch <billy@chainguard.dev>