Use TTY output for errors.
This lets us use user TTYs to output errors / panics directly to the
user instead of them getting lost to stderr. Unfortunately if there is
not a TTY available (this is common for CI runners - incl GitHub Actions
by default_ then the GITSIGN_LOG variable is still required for debug
output.

Signed-off-by: Billy Lynch <billy@chainguard.dev>