Calculate correct SHA for signed Tags.
Previously we were trying to marshal tags into commit objects, which
go-git was happily doing, but ignoring non-matching fields. This change
tries to detect whether we are signing a commit or tag and encode the
matching type.

Also updates README for more copy/paste instructions for signing tags.

BREAKING CHANGE: Since this is fixing how the tag SHA was meant to be
calculated, this breaks the rekor entry lookup for older versions that
use the incorrect behavior. Those tags will be considered unverified
unless they are resigned by a newer version of gitsign: `git tag -f -s <tag name>`

Signed-off-by: Billy Lynch <billy@chainguard.dev>