Add gitsign-attest
This is an experimental binary to play around with Git attestations. The
goal is to ultimately add this to the main gitsign binary, but by having
as separate binary this lets us play around with things without
distributing to users. (also I think the main binary needs some work to
support subcommands).

Signed-off-by: Billy Lynch <billy@chainguard.dev>