Remove usage of getopt to fix release.
There seems to be some conflict with goreleaser, getopt/v2, getopt/v1
(through an indirect dep). We wanted to move off of getopt anyway, so
this is just a convenient excuse to make the change.

Verified we can get past the build step with

```sh
goreleaser release --skip-publish --rm-dist --debug --snapshot
```

so hopefully this is the last of it.

Signed-off-by: Billy Lynch <billy@chainguard.dev>