Add gitsign verify.
This adds a new subcommand that mirrors cosign verify for certificate
claim verification.

Previously we relied on `git verify-commit` for commit verification.
While this did check that the signature was valid and it exists in
rekor, it did not check whether the identity was what was expected,
because Git does not give controls over this via the commit signature
interface.

This command provides this functionality.

Also adds a warning to the output of the `git verify-commit` output
warning users that that verification mechanism may not be complete.

Signed-off-by: Billy Lynch <billy@chainguard.dev>