Fix offline verification marshalling, add e2e tests.
e2e verification was failing because we were using the wrong body to
recreate the tlog message.

To test this, this also adds an e2e test that verifies a real commit
only using offline verification. This is a bit brittle at the moment
because it will mutate the local TUF client state, but this will
hopefully be improved once the TUF client changes land in
sigstore/client-go.

Signed-off-by: Billy Lynch <billy@chainguard.dev>