Run e2e Go tests first.
I suspect why the e2e tests are failing is because we're running this
after the staging tests when the sigstore root has been changed to
staging.

Current the tests hardcode prod sigstore, which is likely what is causing the issues
since we're trying to verify a prod signature with the staging root.
TBD if this is the right thing to do (or if we should use staging instead),
but this should hopefully fix us for the time being.