attest: force https for rekor client
For some reason this defaults to http, which results in the server
terminating the request (my guess is this was previously redirecting
http->https before). This changes forces the https scheme (hopefully
this isn't disruptive for non-public good instances, but we can revisit
other changes later if we need to).

Signed-off-by: Billy Lynch <billy@chainguard.dev>