Modernize attest statements (#620)
* Modernize attest statements

attest no longer uses the cosign library to generate the Statement
used in the attestation.

This lets us:

1. Use `gitCommit` as the digest type (which is what we were doing anyways)
2. Use the correct `_type` value for modern in-toto statements

In the future it will let us add annotations on the subject.

Note that this also changes the user behavior in that it:
1. Changes the things mentioned above.
2. Doesn't let users use the convenience attestation types from the command
   line.  They must instead specify the full predicate type.

refs #611

Signed-off-by: Tom Hennen <tomhennen@google.com>

* gofmt

Signed-off-by: Tom Hennen <tomhennen@google.com>

* Restructure test to allow deep comparison of the JSON

The prior setup needed the contents to be exactly equal.
Unfortunately protojson explicitly doesn't maintain
byte-for-byte equality. To address that the test now
deeply examines the 'attestation' (.sig) file produced
parses the payload JSON into a struct and then compares
that.

This provides somewhat easier to understand error diffs
when things go wrong, but most importantly it appears
stable even if field order or spacing change.

Signed-off-by: Tom Hennen <tomhennen@google.com>

* remove byte conversion to make linter happy

Signed-off-by: Tom Hennen <tomhennen@google.com>

* use gitCommit/gitTree as appropriate

Signed-off-by: Tom Hennen <tomhennen@google.com>

* remove old provenance template

Signed-off-by: Tom Hennen <tomhennen@google.com>

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>