Fix e2e tests timing issues
We currently have a timing issue with the provisioned OIDC token when we
run all tests as part of the e2e suite. This updates it so only the e2e
tests are run. Also, gitsign install is moved up so all the tests using
the token are run in quick succession after the token is provisioned.

Signed-off-by: Aditya Sirish <aditya@saky.in>